change response of product API

+++ app/Http/Controllers/ProductController.php
@@ -32,7 +32,7 @@ class ProductController extends Controller
                 }
             }
 
-            $value->price = filter_var($value->price, FILTER_SANITIZE_NUMBER_INT);
+            $value->price = (int)filter_var($value->price, FILTER_SANITIZE_NUMBER_INT);
 
             // $value = (object) array_merge( (array)$value, array( 'list_picture' => $list_picture ) );
             $value->list_picture = $list_picture;
